move around code to match structure
