+++ src/serlo-editor-integration/plugins.tsx
@@ -23,7 +23,7 @@ import { equationsPlugin } from '@/serlo-editor/plugins/equations'
 import { errorPlugin } from '@/serlo-editor/plugins/error'
 import { exercisePlugin } from '@/serlo-editor/plugins/exercise'
 import { geoGebraPlugin } from '@/serlo-editor/plugins/geogebra'
-import { H5pPlugin } from '@/serlo-editor/plugins/h5p/h5p'
+import { H5pPlugin } from '@/serlo-editor/plugins/h5p'
 import { createHighlightPlugin } from '@/serlo-editor/plugins/highlight'
 import { imagePlugin } from '@/serlo-editor/plugins/image/image-with-serlo-config'
 import { injectionPlugin } from '@/serlo-editor/plugins/injection'
